Move the following scripts out of their folders in lib/entities: The black knight, green knight, mammoth, spikeball trap and turret Remove blackknight library from main.lua
