+++ lib/demo.lua
@@ -1,6 +1,6 @@
 local module = {}
 
-module.DEMO_MAX_WORLD = 1
+module.DEMO_MAX_WORLD = 6
 module.DEMO_TUTORIAL_AVAILABLE = false
 
 return module
