+++ lib/entities/black_knight.lua
