+++ lib/entities/laser_turrent.lua
