+++ lib/entities/mammoth.lua
