+++ lib/entities/spikeball_trap.lua
